Move annotations.idmanytoone tests and fix property lookup of identifierMapper
