+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/BasketItemsPK.java
@@ -6,7 +6,7 @@
  */
 
 //$
-package org.hibernate.test.annotations.idmanytoone;
+package org.hibernate.orm.test.annotations.idmanytoone;
 import java.io.Serializable;
 import javax.persistence.Basic;
 import javax.persistence.CascadeType;
