+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/Customers.java
@@ -6,7 +6,7 @@
  */
 
 //$
-package org.hibernate.test.annotations.idmanytoone;
+package org.hibernate.orm.test.annotations.idmanytoone;
 import java.io.Serializable;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
