+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/alphabetical/A.java
@@ -1,0 +1,45 @@
+/*
+ * Hibernate, Relational Persistence for Idiomatic Java
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+
+//$
+package org.hibernate.orm.test.annotations.idmanytoone.alphabetical;
+import java.util.List;
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+
+@Entity
+public class A {
+
+	@Id
+	private int id;
+
+	@OneToMany( mappedBy = "parent" )
+	List<C> children;
+
+	public A() {
+	}
+
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public List<C> getChildren() {
+		return children;
+	}
+
+	public void setChildren(List<C> children) {
+		this.children = children;
+	}
+
+
+}
+
