+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/alphabetical/Acces.java
@@ -5,27 +5,19 @@
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
 
-//$Id$
-package org.hibernate.test.annotations.idmanytoone;
+//$
+package org.hibernate.orm.test.annotations.idmanytoone.alphabetical;
+import java.io.Serializable;
+import java.math.BigInteger;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
+import javax.persistence.ManyToOne;
 
-/**
- * @author Emmanuel Bernard
- */
 @Entity
-public class Project {
-
+public class Acces implements Serializable {
 	@Id
-	@GeneratedValue
-	private int id;
+	private BigInteger idpk;
 
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
+	@ManyToOne
+	private Droitacces idpkdracc;
 }
