+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/alphabetical/AlphabeticalIdManyToOneTest.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later.
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.test.annotations.idmanytoone.alphabetical;
+package org.hibernate.orm.test.annotations.idmanytoone.alphabetical;
 
 import org.junit.Test;
