+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/alphabetical/C.java
@@ -6,29 +6,34 @@
  */
 
 //$
-package org.hibernate.test.annotations.idmanytoone.alphabetical;
+package org.hibernate.orm.test.annotations.idmanytoone.alphabetical;
+import java.util.List;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 import javax.persistence.IdClass;
+import javax.persistence.OneToMany;
 
 @Entity
-@IdClass( BId.class )
-public class B {
+@IdClass( CId.class )
+public class C {
 
 	@Id
-	private C parent;
+	private A parent;
 
 	@Id
 	private int sequenceNumber;
 
-	public B() {
+	@OneToMany( mappedBy = "parent" )
+	List<B> children;
+
+	public C() {
 	}
 
-	public C getParent() {
+	public A getParent() {
 		return parent;
 	}
 
-	public void setParent(C parent) {
+	public void setParent(A parent) {
 		this.parent = parent;
 	}
 
@@ -38,6 +43,14 @@ public class B {
 
 	public void setSequenceNumber(int sequenceNumber) {
 		this.sequenceNumber = sequenceNumber;
+	}
+
+	public List<B> getChildren() {
+		return children;
+	}
+
+	public void setChildren(List<B> children) {
+		this.children = children;
 	}
