+++ hibernate-core/src/test/java/org/hibernate/orm/test/annotations/idmanytoone/alphabetical/Service.java
@@ -6,17 +6,14 @@
  */
 
 //$
-package org.hibernate.test.annotations.idmanytoone.alphabetical;
+package org.hibernate.orm.test.annotations.idmanytoone.alphabetical;
 import java.math.BigInteger;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
+
 
 @Entity
-public class Benefserv {
-	@Id
-	private BigInteger idpk;
-
-	@ManyToOne
-	private Service idpkser;
+public class Service {
+    @Id
+    private BigInteger idpk;
 }
